# docs/swagger.yaml
basePath: /
definitions:
...
  event.CreateRequest:
    properties:
      category:
        example: festas_shows
        type: string
      description:
        type: string
      endsAt:
        type: string
      location:
        $ref: '#/definitions/event.LocationRequest'
      name:
        example: Festival Aurora
        type: string
      salesMode:
        description: |-
          SalesMode is how this event sells: `counted` by the number, the way a
          party does, or `seated` by the chair, with a row and a seat number. It is
          a declaration and not a fact about inventory, the seats themselves
          answer whether a night has any, but nothing else can be derived from an
          event that has no tiers yet, and the interface has to know which question
          to ask next. Empty means counted.
        enum:
        - counted
        - seated
        example: counted
        type: string
      startsAt:
        example: "2026-11-15T22:00:00-03:00"
        type: string
      status:
        enum:
        - draft
        - published
        - cancelled
        example: draft
        type: string
    type: object
  event.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  event.EventEnvelope:
    properties:
      data:
        $ref: '#/definitions/event.EventResponse'
    type: object
...
  event.ListEnvelope:
    properties:
      data:
        items:
          $ref: '#/definitions/event.ListingResponse'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  event.ListingResponse:
    properties:
      availableTickets:
        type: integer
      category:
        type: string
      createdAt:
        type: string
      description:
        type: string
      endsAt:
        type: string
      fromPriceCents:
        description: |-
          FromPriceCents is the cheapest tier still on sale. Null renders as
          "Esgotado" rather than as "free".
        type: integer
      id:
        example: evt_a1b2c3d4
        type: string
      location:
        $ref: '#/definitions/event.LocationResponse'
      media:
        items:
          $ref: '#/definitions/event.MediaResponse'
        type: array
      name:
        type: string
      salesMode:
        enum:
        - counted
        - seated
        type: string
      slug:
        example: festival-aurora
        type: string
      startsAt:
        type: string
      status:
        type: string
      updatedAt:
        type: string
    type: object
...
  event.MediaResponse:
    properties:
      blurDataUrl:
        type: string
      contentType:
        type: string
      height:
        type: integer
      id:
        type: string
      kind:
        enum:
        - image
        - video
        type: string
      position:
        type: integer
      sizeBytes:
        type: integer
      url:
        type: string
      width:
        description: |-
          Width, Height and BlurDataURL are what a client needs to reserve the
          right box and paint something before the bytes arrive. Zero and empty
          when the asset predates the pipeline that generates them.
        type: integer
    type: object
  event.PinRequest:
    properties:
      latitude:
        example: -3.807
        type: number
      longitude:
        example: -38.522
        type: number
    type: object
  event.TierListEnvelope:
    properties:
      data:
        items:
          $ref: '#/definitions/event.TierResponse'
        type: array
    type: object
...
  event.UpdateRequest:
    properties:
      category:
        example: festas_shows
        type: string
      description:
        type: string
      endsAt:
        type: string
      location:
        $ref: '#/definitions/event.LocationRequest'
      name:
        example: Festival Aurora
        type: string
      salesMode:
        description: |-
          SalesMode is how this event sells: `counted` by the number, the way a
          party does, or `seated` by the chair, with a row and a seat number. It is
          a declaration and not a fact about inventory, the seats themselves
          answer whether a night has any, but nothing else can be derived from an
          event that has no tiers yet, and the interface has to know which question
          to ask next. Empty means counted.
        enum:
        - counted
        - seated
        example: counted
        type: string
      startsAt:
        example: "2026-11-15T22:00:00-03:00"
        type: string
      status:
        enum:
        - draft
        - published
        - cancelled
        example: draft
        type: string
    type: object
...
info:
  contact: {}
  description: API base para autenticação e gestão de tickets, organizada com Clean
    Architecture.
  title: Vozkot Tickets API
  version: "1.0"
paths:
...
  /api/v1/events:
    get:
      description: Lista os eventos do operador autenticado, incluindo rascunhos.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/event.ListEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/event.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar meus eventos
      tags:
      - Eventos
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados do evento
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/event.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/event.EventEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/event.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/event.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Criar um evento
      tags:
      - Eventos
  /api/v1/events/{id}:
    delete:
      description: Recusa com 409 se o evento ainda tiver lotes de ingressos, porque
        esses lotes podem ter pedidos e os pedidos são o registro do dinheiro que
        mudou de mãos.
      parameters:
      - description: ID do evento
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Removido
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/event.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Excluir um evento
      tags:
      - Eventos
    get:
      parameters:
      - description: ID do evento
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/event.EventEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/event.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Consultar um evento
      tags:
      - Eventos
    put:
      consumes:
      - application/json
      parameters:
      - description: ID do evento
        in: path
        name: id
        required: true
        type: string
      - description: Dados do evento
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/event.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/event.EventEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/event.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/event.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/event.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Atualizar um evento
      tags:
      - Eventos
...
  /api/v1/events/{id}/media:
    post:
      consumes:
      - multipart/form-data
      description: Anexa uma imagem ou vídeo ao evento, um arquivo por requisição.
      parameters:
      - description: ID do evento
        in: path
        name: id
        required: true
        type: string
      - description: Arquivo
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/event.MediaResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/event.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/event.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Enviar mídia do evento
      tags:
      - Eventos
  /api/v1/events/{id}/media/{mediaId}:
    delete:
      parameters:
      - description: ID do evento
        in: path
        name: id
        required: true
        type: string
      - description: ID da mídia
        in: path
        name: mediaId
        required: true
        type: string
      responses:
        "204":
          description: Removido
      security:
      - BearerAuth: []
      summary: Remover mídia do evento
      tags:
      - Eventos
  /api/v1/events/{id}/pin:
    post:
      consumes:
      - application/json
      description: 'Grava as coordenadas escolhidas por uma pessoa. Elas valem mais
        que qualquer geocodificação: uma edição posterior no endereço não as sobrescreve.'
      parameters:
      - description: ID do evento
        in: path
        name: id
        required: true
        type: string
      - description: Coordenadas
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/event.PinRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/event.EventEnvelope'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/event.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Posicionar o pino no mapa
      tags:
      - Eventos
  /api/v1/events/{id}/publish:
    post:
      parameters:
      - description: ID do evento
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/event.EventEnvelope'
      security:
      - BearerAuth: []
      summary: Publicar um evento
      tags:
      - Eventos
...
  /api/v1/events/{id}/unpublish:
    post:
      parameters:
      - description: ID do evento
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/event.EventEnvelope'
      security:
      - BearerAuth: []
      summary: Despublicar um evento
      tags:
      - Eventos
...
  /api/v1/public/events:
    get:
      description: A vitrine pública. Aceita busca textual, categoria, cidade, intervalo
        de datas, preço máximo, apenas gratuitos, apenas com ingressos disponíveis,
        ordenação e paginação. Só retorna eventos publicados, independentemente do
        que o cliente enviar.
      parameters:
      - description: Busca por nome, local, cidade ou descrição
        in: query
        name: q
        type: string
      - description: Categoria
        in: query
        name: category
        type: string
      - description: Cidade
        in: query
        name: city
        type: string
      - description: Eventos a partir desta data (RFC3339)
        in: query
        name: from
        type: string
      - description: Eventos até esta data (RFC3339)
        in: query
        name: until
        type: string
      - description: Preço máximo em centavos do lote mais barato
        in: query
        name: maxPrice
        type: integer
      - description: Apenas eventos gratuitos
        in: query
        name: free
        type: boolean
      - description: Apenas eventos com ingressos disponíveis
        in: query
        name: available
        type: boolean
      - description: Ordenação
        enum:
        - relevance
        - starts_at
        - price
        - created_at
        in: query
        name: sort
        type: string
      - description: Itens por página (padrão 24, máximo 60)
        in: query
        name: limit
        type: integer
      - description: Deslocamento da paginação
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/event.ListEnvelope'
      summary: Listar eventos publicados
      tags:
      - Catálogo
  /api/v1/public/events/{id}/tiers:
    get:
      description: 'Os lotes que o evento vende, do mais barato ao mais caro. Só retorna
        lotes à venda de eventos publicados: é o que o painel de compra da página
        do evento monta.'
      parameters:
      - description: ID do evento
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/event.TierListEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/event.ErrorResponse'
      summary: Lotes de ingressos à venda
      tags:
      - Catálogo
  /api/v1/public/events/{slug}:
    get:
      description: 'A página pública do evento, resolvida pelo slug da URL. Um rascunho
        responde 404 e não 403: um 403 confirmaria que o evento existe naquele endereço.'
      parameters:
      - description: Slug do evento
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/event.EventEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/event.ErrorResponse'
      summary: Consultar um evento publicado
      tags:
      - Catálogo
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Use o formato: Bearer {token}. Navegadores também podem autenticar
      por cookies httpOnly.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"